linter fix and refactoring
