+++ src/main.js
@@ -1,6 +1,6 @@
 import FilterView from './view/filter-view.js';
-import SortingView from "./view/sorting-view.js";
-import EventsListPresenter from "./presenter/events-list-presenter.js";
+import SortingView from './view/sorting-view.js';
+import EventsListPresenter from './presenter/events-list-presenter.js';
 import { render } from './render.js';
 
 const filterContainerElement = document.querySelector('.trip-controls__filters');
