+++ src/presenter/events-list-presenter.js
@@ -4,18 +4,18 @@ import EventView from '../view/event-in-list-view';
 import { render } from '../render';
 
 export default class EventsListPresenter {
-    eventsListComponent = new EventsListView();
+  eventsListComponent = new EventsListView();
 
-    constructor(eventsListSection) {
-        this.eventsListSection = eventsListSection;
+  constructor(eventsListSection) {
+    this.eventsListSection = eventsListSection;
+  }
+
+  init() {
+    render(this.eventsListComponent, this.eventsListSection);
+    render(new EventEditView(), this.eventsListComponent.getElement());
+
+    for (let i = 0; i < 3; i++) {
+      render(new EventView(), this.eventsListComponent.getElement());
     }
-
-    init() {
-        render(this.eventsListComponent, this.eventsListSection);
-        render(new EventEditView(), this.eventsListComponent.getElement());
-
-        for (let i = 0; i < 3; i++) {
-            render(new EventView(), this.eventsListComponent.getElement());
-        }
-    }
+  }
 }
