+++ src/template/events-list-template.js
@@ -1,0 +1,5 @@
+function createEventsListTemplate() {
+  return '<ul class="trip-events__list"></ul>';
+}
+
+export {createEventsListTemplate};
