+++ src/view/sorting-view.js
@@ -1,50 +1,20 @@
-import { createElement } from "../render";
-
-function createSortingTemplate() {
-    return (
-        `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
-            <div class="trip-sort__item  trip-sort__item--day">
-                <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day">
-                <label class="trip-sort__btn" for="sort-day">Day</label>
-            </div>
-            
-            <div class="trip-sort__item  trip-sort__item--event">
-                <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" disabled>
-                <label class="trip-sort__btn" for="sort-event">Event</label>
-            </div>
-            
-            <div class="trip-sort__item  trip-sort__item--time">
-                <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time">
-                <label class="trip-sort__btn" for="sort-time">Time</label>
-            </div>
-            
-            <div class="trip-sort__item  trip-sort__item--price">
-                <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price" checked>
-                <label class="trip-sort__btn" for="sort-price">Price</label>
-            </div>
-            
-            <div class="trip-sort__item  trip-sort__item--offer">
-                <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer" disabled>
-                <label class="trip-sort__btn" for="sort-offer">Offers</label>
-            </div>
-        </form>`
-    );
-}
+import { createElement } from '../render';
+import { createSortingTemplate } from '../template/sorting-template.js';
 
 export default class SortingView {
-    getTemplate() {
-        return createSortingTemplate();
+  getTemplate() {
+    return createSortingTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
     }
 
-    getElement() {
-        if (!this.element) {
-            this.element = createElement(this.getTemplate());
-        }
+    return this.element;
+  }
 
-        return this.element;
-    }
-
-    removeElement() {
-        this.element = null;
-    }
+  removeElement() {
+    this.element = null;
+  }
 }
